License fix: Move detector correction to io/ for MIT licensing
