+++ src/libertem/io/corrections/__init__.py
@@ -1,0 +1,3 @@
+from .corrset import CorrectionSet
+
+__all__ = ["CorrectionSet"]
