+++ src/libertem/io/dataset/base/partition.py
@@ -4,7 +4,7 @@ from typing import Optional
 import numpy as np
 
 from libertem.common import Slice, Shape
-from libertem.corrections import CorrectionSet
+from libertem.io.corrections import CorrectionSet
 from .tiling import DataTile
 from .tiling_scheme import TilingScheme
 from .meta import DataSetMeta
