+++ src/libertem/io/dataset/frms6.py
@@ -13,7 +13,7 @@ import scipy.io as sio
 import numpy as np
 
 from libertem.common.math import prod
-from libertem.corrections import CorrectionSet
+from libertem.io.corrections import CorrectionSet
 from libertem.common import Shape, Slice
 from libertem.web.messages import MessageConverter
 from .base import (
