+++ tests/corrections/test_detector.py
@@ -3,7 +3,7 @@ import sparse
 import pytest
 
 from libertem.utils.generate import gradient_data, exclude_pixels
-from libertem.corrections import detector
+from libertem.io.corrections import detector
 from libertem.masks import is_sparse
